+++ docs/ru/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 2019-03-02
 
+### [v2.3.1][v2.3.1]
+
+* Исправлено: теперь нет проверки транспортных средств других фракций при событии "on_selected_entity_changed", "on_player_driving_changed_state"
+* Изменено: описание
+
 ### [v2.3.0][v2.3.0]
 
 * Добавлено: индивидуальная визуализация
@@ -132,6 +137,7 @@
 
 * Первый релиз для 0.16
 
+[v2.3.1]: https://mods.factorio.com/mod/diplomacy/downloads
 [v2.3.0]: https://mods.factorio.com/download/show-health-and-shield/5c7a6bb4769efa000bf59a64
 [v2.2.0]: https://mods.factorio.com/download/show-health-and-shield/5c7a5b8c769efa000df6000e
 [v2.1.0]: https://mods.factorio.com/download/show-health-and-shield/5c78e511507d61000dcd440d